Allow functions to return custom datatypes
As things stand, the extension only allows for "primitive" Lua types
that man directly onto the SQLite type system (eg. nil, numbers,
strings, booleans) to be passed around.  This change allows for tables
and userdata to be passed around, as long as they have a metatable
registered with the extension.  The metatable must have __tosqlite
and __fromsqlite metafunctions, which describe how to convert to and
from SQLite values.